refactor: isolate server to use it in tests
